requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
alembic>=1.14.0
psycopg2-binary>=2.9.10
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-multipart>=0.0.18
python-dotenv>=1.0.1
pymupdf>=1.25.0
pytesseract>=0.3.13
Pillow>=11.0.0
aiohttp>=3.9.0

streamlit>=1.38.0

pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
